+++ src/components/roster/RosterPage.tsx
@@ -132,7 +132,7 @@ export default function RosterPage() {
         </div>
         {!edit.editing && (
           <Space>
-            <Button icon={<SettingOutlined />} onClick={() => setDrawerOpen(true)}>Manage Roster</Button>
+            <Button icon={<SettingOutlined />} onClick={() => setDrawerOpen(true)}>Manage Shift Patterns</Button>
             <Tooltip title={editableMonth ? undefined : PAST_MONTH_TOOLTIP}>
               <Button icon={<EditOutlined />} disabled={!editableMonth} onClick={edit.start}>
                 Edit Roster